Refactoring NP mapping with a base class (#1287)
